rcll: workpiece-tracking fact, cleaner en/disableing
We might later on, depending on how the real world testings with the
barcodes goes, need to disable the tracking on the fly if some of the
readers failed.  Specially, BS or RS. The new 'workpiece-tracking' fact
has a cleaner way to keep the status of tracking (dis/enabled) and the
reason of that status.